변경 - BoardPicture 변수타입 int에서 long으로 변경

+++ src/main/java/com/example/intermediate/domain/BoardPicture.java
@@ -19,9 +19,9 @@ import javax.validation.constraints.NotNull;
 public class BoardPicture {
     @Id
     @GeneratedValue
-    private Integer ID;
+    private Long ID;
     @NotNull
-    private Integer boardIdx;
+    private Long boardIdx;
 
     // 원본 파일이름과 서버에 저장된 파일 경로를 분리
     // 동일한 이름을 가진 파일이 업로드가 된다면 오류가 생긴다.
